auth implementation, user roles and permissions, controller tests, model updates

+++ app/Http/Controllers/ItemController.php
@@ -6,32 +6,53 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Collection;
 use App\Models\Item;
 use Illuminate\Http\Response;
+use Illuminate\Support\Facades\Auth;
 
 class ItemController extends Controller
 {
     public function index(){
-        return Item::all();
+        if(Auth::user()->role == 'admin' || Auth::user()->role == 'employee'){
+            return Item::all();
+        }
     }
 
     public function read(Item $item){
-        return $item;
+        if(Auth::user()->role == 'admin' || Auth::user()->role == 'employee'){
+            return $item;
+        }
+        else{
+            return response()->json(['message' => 'user not authorized for this action!'], 401);
+        }
     }
 
     public function create(Request $request){
-        $item = Item::create($request->all());
-
-        return response()->json($item, 201);
+        if($request->user()->role == 'admin'){
+            $item = Item::create($request->all());
+            return response()->json($item, 201);
+        }
+        else{
+            return response()->json(['message' => 'user not authorized for this action!'], 401);
+        }
     }
 
     public function update(Request $request, Item $item){
-        $item->update($request->all());
-
-        return response()->json($item, 200);
+        if($request->user()->role == 'admin'){
+            $item->update($request->all());
+            return response()->json($item, 200);
+        }
+        else{
+            return response()->json(['message' => 'user not authorized for this action!'], 401);
+        }
     }
 
     public function delete(Request $request, Item $item){
-        $item->delete();
+        if($request->user()->role == 'admin'){
+            $item->delete();
 
-        return response()->json(null, 204);
+            return response()->json(null, 204);
+        }
+        else{
+            return response()->json(['message' => 'user not authorized for this action!'], 401);
+        }
     }
 }
